codeforces/cpp/1900-1999/E/1971E.cpp

+++ codeforces/cpp/1900-1999/E/1971E.cpp
@@ -50,7 +50,7 @@ void solve()
         if (index < a.size() - 1)
         {
             response += static_cast<int>(trunc(
-                static_cast<float>(value - a[index]) / static_cast<float>(a[index + 1] - a[index]) * static_cast<float>(
+                static_cast<double>(value - a[index]) / static_cast<double>(a[index + 1] - a[index]) * static_cast<double>(
                     b[index + 1] - b[index])));
         }
